Move some code from 'keycloak' pkg to 'tenant'
